Changed seeding to call the NYC Open Data API instead of reading a .json file
Added the soda-ruby gem in order to communicate with the NYC Open Data API. Encapsulated our seeding mechanism in the Adapters::NYCOpenData class and changed the data source to the NYC Open Data API instead of a locally stored .json file. Also switched the order of our migrations so that rolling back will drop the users table last.