Further built out the favorites functionality
Added a conditional to check to see if you have a location added to your favorites before asking if you want to add it to your favorites. Moved the definition of the users location to the initialize method for the WifiRunner class. Added a method for the user to see a list of their favorites, ordered by distance.